backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
pinecone-client==5.0.1
sentence-transformers==3.3.1
google-generativeai==0.8.3
gunicorn==23.0.0